CU-862j5by9q Rename categoryseparation module as category_separation
